requirements.txt
pillow==9.5.0
pandoc-img-glob
panflute